Moves constraints and solvers in a separate folder
