--- src/main/kotlin/ru/hse/plameet/core/Solver.kt
@@ -1,5 +1,0 @@
-package ru.hse.plameet.core
-
-interface Solver {
-    fun solve(events: List<Event>, constraints: List<Constraint>): Schedule
-}
