--- src/main/kotlin/ru/hse/plameet/core/TimetableConstraint.kt
@@ -1,4 +1,0 @@
-package ru.hse.plameet.core
-
-open class TimetableConstraint(slots: List<TimeStamp>, duration: Duration) :
-    SlotsConstraint(TimeSlots(slots.map { it.during(duration) }))
