+++ src/main/kotlin/ru/hse/plameet/core/constraints/AvailabilityConstraint.kt
@@ -1,4 +1,6 @@
-package ru.hse.plameet.core
+package ru.hse.plameet.core.constraints
+
+import ru.hse.plameet.core.*
 
 class AvailabilityConstraint(av: Map<User, List<TimeRange>>) : RequiredConstraint {
