+++ src/test/kotlin/ru/hse/plameet/core/constraints/AvailabilityConstraintTest.kt
@@ -1,10 +1,11 @@
-package ru.hse.plameet.core
+package ru.hse.plameet.core.constraints
 
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import ru.hse.plameet.core.*
 import java.util.stream.Stream
 
 object AvailabilityConstraintTest {
